+++ flycomp/src/lib.rs
@@ -10,9 +10,6 @@ pub mod parse_man;
 
 pub use parse_help::{parse_help, parse_help_argparse, parse_help_clap, parse_help_generic};
 
-#[cfg(test)]
-pub(crate) use parse_help::parse_flag_tokens;
-
 #[derive(Clone, Debug, PartialEq, Eq, Default)]
 pub enum SynthesisStrategy {
     #[default]
@@ -216,23 +213,21 @@ pub fn synthesize_completion<F>(
 where
     F: Fn(&[&str]) -> anyhow::Result<String>,
 {
-    synthesize_completion_with(command_path, &help_runner, &load_manpage_command, strategy)
+    synthesize_completion_with(command_path, &help_runner, strategy)
 }
 
-fn synthesize_completion_with<F, G>(
+fn synthesize_completion_with<F>(
     command_path: &str,
     help_runner: &F,
-    manpage_loader: &G,
     strategy: SynthesisStrategy,
 ) -> anyhow::Result<Command>
 where
     F: Fn(&[&str]) -> anyhow::Result<String>,
-    G: Fn(&str) -> anyhow::Result<Command>,
 {
     match strategy {
         SynthesisStrategy::RunHelp => synthesize_from_help(command_path, help_runner),
-        SynthesisStrategy::ManPage => manpage_loader(command_path),
-        SynthesisStrategy::ManPageThenRunHelp => match manpage_loader(command_path) {
+        SynthesisStrategy::ManPage => load_manpage_command(command_path),
+        SynthesisStrategy::ManPageThenRunHelp => match load_manpage_command(command_path) {
             Ok(command) => Ok(command),
             Err(error) => {
                 log::debug!(
@@ -246,8 +241,7 @@ where
     }
 }
 
-// TODO: we should have kept track of the command name insteaf of parsing it from the file name.
-fn command_name(command_path: &str) -> String {
+fn command_basename(command_path: &str) -> String {
     std::path::Path::new(command_path)
         .file_name()
         .and_then(|s| s.to_str())
@@ -256,7 +250,7 @@ fn command_name(command_path: &str) -> String {
 }
 
 fn load_manpage_command(command_path: &str) -> anyhow::Result<Command> {
-    let cmd_name = command_name(command_path);
+    let cmd_name = command_basename(command_path);
     let manpage_path = locate_manpage(&cmd_name)?;
     let manpage_content = read_manpage_source(&manpage_path)?;
 
@@ -318,7 +312,7 @@ where
     let mut root = parse_help(&top_help);
 
     // Always use the basename of the supplied path as the canonical name.
-    let cmd_name = command_name(command_path);
+    let cmd_name = command_basename(command_path);
     root.name = Some(cmd_name);
 
     // ── iterative subcommand exploration ─────────────────────────────────────
@@ -423,7 +417,7 @@ pub fn generate_completion_script(
 ) -> anyhow::Result<String> {
     let parsed_cmd =
         synthesize_completion(command_path, |args| run_help(command_path, args), strategy)?;
-    let cmd_name = command_name(command_path);
+    let cmd_name = command_basename(command_path);
 
     let mut clap_cmd = to_clap_command(&parsed_cmd);
     let mut output = Vec::new();
@@ -438,783 +432,6 @@ pub fn generate_completion_script(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::cell::Cell;
-
-    // ── helpers ──────────────────────────────────────────────────────────────
-
-    /// Returns the long names of all args in `cmd`.
-    fn long_names(cmd: &Command) -> Vec<&str> {
-        cmd.args.iter().filter_map(|a| a.long.as_deref()).collect()
-    }
-
-    /// Returns the short names of all args in `cmd`.
-    fn short_names(cmd: &Command) -> Vec<&str> {
-        cmd.args.iter().filter_map(|a| a.short.as_deref()).collect()
-    }
-
-    /// Returns the names of all sub-commands in `cmd`.
-    fn subcommand_names(cmd: &Command) -> Vec<&str> {
-        cmd.subcommands
-            .iter()
-            .filter_map(|s| s.name.as_deref())
-            .collect()
-    }
-
-    /// Looks up an arg by its long name.
-    fn arg_by_long<'a>(cmd: &'a Command, long: &str) -> Option<&'a Arg> {
-        cmd.args.iter().find(|a| a.long.as_deref() == Some(long))
-    }
-
-    /// Looks up a subcommand by name.
-    fn subcommand_by_name<'a>(cmd: &'a Command, name: &str) -> Option<&'a Command> {
-        cmd.subcommands
-            .iter()
-            .find(|s| s.name.as_deref() == Some(name))
-    }
-
-    // ── flyline --help (clap) ────────────────────────────────────────────────
-
-    const FLYLINE_HELP: &str = r#"Usage: flyline [OPTIONS] [COMMAND]
-
-Commands:
-  set-agent-mode        Configure AI agent mode.
-  create-prompt-widget  Create a custom prompt widget.
-  set-style             Configure the colour palette.
-  set-cursor            Configure the cursor appearance and animation.
-  key                   Manage keybindings.
-  dump-logs             Dump in-memory logs to file.
-  stream-logs           Dump current logs to PATH and append new logs.
-  run-tutorial          Run the interactive tutorial for first-time users.
-  help                  Print this message or the help of the given subcommand(s)
-
-Options:
-      --version
-          Show version information
-
-      --log-level <LEVEL>
-          Set the logging level
-          
-          [possible values: error, warn, info, debug, trace]
-
-      --load-zsh-history [<PATH>]
-          Load Zsh history in addition to Bash history. Optionally specify a PATH to the Zsh history file; if omitted, defaults to $HOME/.zsh_history
-
-      --show-animations [<SHOW_ANIMATIONS>]
-          Show animations
-          
-          [possible values: true, false]
-
-      --show-inline-history [<SHOW_INLINE_HISTORY>]
-          Show inline history suggestions
-          
-          [possible values: true, false]
-
-      --auto-close-chars [<AUTO_CLOSE_CHARS>]
-          Enable automatic closing character insertion (e.g. insert `)` after `(`)
-          
-          [possible values: true, false]
-
-      --matrix-animation [<MATRIX_ANIMATION>]
-          Run matrix animation in the terminal background. Use `on` to always show it, `off` to disable it, or an integer number of seconds to show it after that many seconds of inactivity (no keypress or mouse event). Defaults to `off`; passing the flag without a value is equivalent to `on`
-
-      --set-frame-rate <FPS>
-          Render frame rate in frames per second (1–120, default 24)
-
-      --set-mouse-mode <MODE>
-          Mouse capture mode (disabled, simple, smart). Default is smart.
-
-      --send-shell-integration-codes [<SEND_SHELL_INTEGRATION_CODES>]
-          Send shell integration escape codes (OSC 133 / OSC 633): none, only-prompt-pos, or full
-
-      --enable-extended-key-codes [<ENABLE_EXTENDED_KEY_CODES>]
-          Whether to request the use of extended (kitty-protocol) keyboard codes during startup. Enabled by default; pass `--enable-extended-key-codes false` (or with no value) to disable it on terminals that misbehave when the request is sent
-          
-          [possible values: true, false]
-
-  -h, --help
-          Print help (see a summary with '-h')
-
-Read more at https://github.com/HalFrgrd/flyline
-"#;
-
-    #[test]
-    fn test_flyline_help_clap() {
-        let cmd = parse_help(FLYLINE_HELP);
-        assert_eq!(cmd.name.as_deref(), Some("flyline"));
-
-        // Subcommands and their descriptions.
-        let subs = subcommand_names(&cmd);
-        assert!(subs.contains(&"set-agent-mode"));
-        assert!(subs.contains(&"create-prompt-widget"));
-        assert!(subs.contains(&"set-style"));
-        assert!(subs.contains(&"key"));
-        assert!(subs.contains(&"dump-logs"));
-        assert!(subs.contains(&"stream-logs"));
-        assert!(subs.contains(&"run-tutorial"));
-        assert_eq!(
-            subcommand_by_name(&cmd, "set-agent-mode").and_then(|s| s.description.as_deref()),
-            Some("Configure AI agent mode.")
-        );
-        assert_eq!(
-            subcommand_by_name(&cmd, "dump-logs").and_then(|s| s.description.as_deref()),
-            Some("Dump in-memory logs to file.")
-        );
-        assert_eq!(
-            subcommand_by_name(&cmd, "stream-logs").and_then(|s| s.description.as_deref()),
-            Some("Dump current logs to PATH and append new logs.")
-        );
-        assert_eq!(
-            subcommand_by_name(&cmd, "run-tutorial").and_then(|s| s.description.as_deref()),
-            Some("Run the interactive tutorial for first-time users.")
-        );
-
-        // Flag names.
-        let longs = long_names(&cmd);
-        assert!(longs.contains(&"--version"));
-        assert!(
-            !longs.contains(&"--dump-logs"),
-            "--dump-logs should be a subcommand now"
-        );
-        assert!(
-            !longs.contains(&"--stream-logs"),
-            "--stream-logs should be a subcommand now"
-        );
-        assert!(
-            !longs.contains(&"--run-tutorial"),
-            "--run-tutorial should be a subcommand now"
-        );
-        assert!(longs.contains(&"--log-level"));
-        assert!(longs.contains(&"--set-frame-rate"));
-        assert!(longs.contains(&"--set-mouse-mode"));
-        assert!(longs.contains(&"--help"));
-        let shorts = short_names(&cmd);
-        assert!(shorts.contains(&"-h"));
-
-        // Value types (meta-variable hints).
-        assert_eq!(
-            arg_by_long(&cmd, "--log-level").and_then(|a| a.value_type.as_deref()),
-            Some("LEVEL")
-        );
-        assert_eq!(
-            arg_by_long(&cmd, "--set-frame-rate").and_then(|a| a.value_type.as_deref()),
-            Some("FPS")
-        );
-        assert_eq!(
-            arg_by_long(&cmd, "--set-mouse-mode").and_then(|a| a.value_type.as_deref()),
-            Some("MODE")
-        );
-
-        // Descriptions on continuation lines.
-        assert_eq!(
-            arg_by_long(&cmd, "--version").and_then(|a| a.description.as_deref()),
-            Some("Show version information")
-        );
-        assert!(
-            arg_by_long(&cmd, "--log-level")
-                .and_then(|a| a.description.as_deref())
-                .unwrap_or("")
-                .contains("logging level"),
-            "description should mention 'logging level'"
-        );
-    }
-
-    // ── flyline create-anim --help (clap) ────────────────────────────────────
-    const FLYLINE_CREATE_ANIM_HELP: &str = r#"Create a custom prompt animation.
-
-Instances of NAME in prompt strings (PS1, RPS1, PS1_FILL) are replaced
-with the current animation frame on every render.  Frames may include
-ANSI colour sequences written as `\e` (e.g. `\e[33m`).
-
-Examples:
-  flyline create-anim --name "MY_ANIMATION" --fps 10  ⣾ ⣷ ⣯ ⣟ ⡿ ⢿ ⣻ ⣽
-  flyline create-anim --name "john" --ping-pong --fps 5  '\e[33m\u' '\e[31m\u' '\e[35m\u' '\e[36m\u'
-
-Usage: flyline create-anim [OPTIONS] --name <NAME> [FRAMES]...
-
-Arguments:
-  [FRAMES]...
-          One or more animation frames (positional).  Use `\e` for the ESC character
-
-Options:
-      --name <NAME>
-          Name to embed in prompt strings as the animation placeholder
-
-      --fps <FPS>
-          Playback speed in frames per second (default: 10)
-          
-          [default: 10]
-
-      --ping-pong
-          Reverse direction at each end instead of wrapping (ping-pong / bounce mode)
-
-  -h, --help
-          Print help (see a summary with '-h')
-"#;
-
-    #[test]
-    fn test_flyline_create_anim_help_clap() {
-        let cmd = parse_help(FLYLINE_CREATE_ANIM_HELP);
-        assert_eq!(cmd.name.as_deref(), Some("flyline"));
-        assert!(
-            cmd.description
-                .as_deref()
-                .unwrap_or("")
-                .contains("animation"),
-            "description should mention animation"
-        );
-
-        let longs = long_names(&cmd);
-        assert!(longs.contains(&"--name"));
-        assert!(longs.contains(&"--fps"));
-        assert!(longs.contains(&"--ping-pong"));
-        assert!(longs.contains(&"--help"));
-        let shorts = short_names(&cmd);
-        assert!(shorts.contains(&"-h"));
-
-        // Value types.
-        assert_eq!(
-            arg_by_long(&cmd, "--name").and_then(|a| a.value_type.as_deref()),
-            Some("NAME")
-        );
-        assert_eq!(
-            arg_by_long(&cmd, "--fps").and_then(|a| a.value_type.as_deref()),
-            Some("FPS")
-        );
-
-        // Descriptions on continuation lines.
-        assert!(
-            arg_by_long(&cmd, "--name")
-                .and_then(|a| a.description.as_deref())
-                .unwrap_or("")
-                .contains("animation placeholder"),
-            "description should mention 'animation placeholder'"
-        );
-        assert!(
-            arg_by_long(&cmd, "--fps")
-                .and_then(|a| a.description.as_deref())
-                .unwrap_or("")
-                .contains("frames per second"),
-            "description should mention 'frames per second'"
-        );
-        assert!(
-            arg_by_long(&cmd, "--ping-pong")
-                .and_then(|a| a.description.as_deref())
-                .unwrap_or("")
-                .contains("Reverse direction"),
-            "description should mention 'Reverse direction'"
-        );
-    }
-
-    // ── cargo --help (clap) ──────────────────────────────────────────────────
-
-    const CARGO_HELP: &str = r#"Rust's package manager
-
-Usage: cargo [+toolchain] [OPTIONS] [COMMAND]
-       cargo [+toolchain] [OPTIONS] -Zscript [ARGS]...
-
-Options:
-  -V, --version              Print version info and exit
-      --list                 List installed commands
-      --explain <CODE>       Run `rustc --explain CODE`
-  -v, --verbose              Use verbose output (-vv very verbose/build.rs output)
-  -q, --quiet                Do not print cargo log messages
-      --color <WHEN>         Coloring: auto, always, never
-  -C <DIRECTORY>             Change to DIRECTORY before doing anything (nightly-only)
-      --frozen               Require Cargo.lock and cache are up to date
-      --locked               Require Cargo.lock is up to date
-      --offline              Run without accessing the network
-      --config <KEY=VALUE>   Override a configuration value
-  -Z <FLAG>                  Unstable (nightly-only) flags to Cargo, see 'cargo -Z help' for details
-  -h, --help                 Print help
-
-Some common cargo commands are (see all commands with --list):
-    build, b    Compile the current package
-    check, c    Analyze the current package and report errors, but don't build object files
-    clean       Remove the generated artifacts
-    doc, d      Build this package's and its dependencies' documentation
-    run, r      Run a binary or example of the local package
-    test, t     Run the tests
-    bench       Run the benchmarks
-    update      Update dependencies listed in Cargo.lock
-    search      Search registry for crates
-    publish     Publish this package to the registry
-    add         Add dependencies to a manifest file
-    remove      Remove dependencies from a manifest file
-
-See 'cargo help <command>' for more information on a specific command.
-"#;
-
-    #[test]
-    fn test_cargo_help_clap() {
-        let cmd = parse_help(CARGO_HELP);
-        assert_eq!(cmd.name.as_deref(), Some("cargo"));
-
-        let longs = long_names(&cmd);
-        assert!(longs.contains(&"--version"));
-        assert!(longs.contains(&"--verbose"));
-        assert!(longs.contains(&"--quiet"));
-        assert!(longs.contains(&"--explain"));
-        assert!(longs.contains(&"--color"));
-        assert!(longs.contains(&"--help"));
-        let shorts = short_names(&cmd);
-        assert!(shorts.contains(&"-V"));
-        assert!(shorts.contains(&"-v"));
-        assert!(shorts.contains(&"-h"));
-
-        // Value types inferred from inline meta-variables.
-        assert_eq!(
-            arg_by_long(&cmd, "--explain").and_then(|a| a.value_type.as_deref()),
-            Some("CODE")
-        );
-        assert_eq!(
-            arg_by_long(&cmd, "--color").and_then(|a| a.value_type.as_deref()),
-            Some("WHEN")
-        );
-    }
-
-    // ── python --help (argparse-like, but actually custom) ───────────────────
-
-    const PYTHON_HELP: &str = r#"usage: python3 [option] ... [-c cmd | -m mod | file | -] [arg] ...
-Options and arguments (and corresponding environment variables):
--b     : issue warnings about str(bytes_instance), str(bytearray_instance)
-         and comparing bytes/bytearray with str. (-bb: issue errors)
--B     : don't write .pyc files on import; also PYTHONDONTWRITEBYTECODE=x
--c cmd : program passed in as string (terminates option list)
--d     : debug output from parser; also PYTHONDEBUG=x
--E     : ignore PYTHON* environment variables (such as PYTHONPATH)
--h     : print this help message and exit (also --help)
--i     : inspect interactively after running script; forces a prompt even
-         if stdin does not appear to be a terminal; also PYTHONINSPECT=x
--I     : isolate Python from the user's environment (implies -E and -s)
--m mod : run library module as a script (terminates option list)
--O     : remove assert and __debug__-dependent statements; add .opt-1 before
-         .pyc extension; also PYTHONOPTIMIZE=x
--OO    : do -O changes and also discard docstrings; add .opt-2 before
-         .pyc extension
--q     : don't print version and copyright messages on interactive startup
--s     : don't add user site directory to sys.path; also PYTHONNOUSERSITE=x
--S     : don't imply 'import site' on initialization
--u     : force the stdout and stderr streams to be unbuffered;
-         this option has no effect on stdin; also PYTHONASYNCIOGENIGNORE=x
--v     : verbose (trace import statements); also PYTHONVERBOSE=x;
-         can be supplied multiple times to increase verbosity
--V     : print the Python version number and exit (also --version)
-         --version :  print just the version number to stdout and exit
--W arg : warning control; arg is action:message:category:module:lineno
-         also PYTHONWARNINGS=x
--x     : skip first line of source, allowing use of non-Unix forms of #!cmd
--X opt : set implementation-specific option
---check-hash-based-pycs always|default|never:
-    control how Python invalidates hash-based .pyc files
-file   : program read from script file
--      : program read from stdin (default; interactive mode if a tty)
-arg ...: arguments passed to program in sys.argv[1:]
-
-Other environment variables:
-PYTHONSTARTUP: file executed on interactive startup (no default)
-PYTHONPATH   : ':'-separated list of directories prefixed to the
-               default module search path.  The result is sys.path.
-"#;
-
-    #[test]
-    fn test_python_help_generic() {
-        // python --help doesn't follow argparse or clap exactly; falls through to generic.
-        let cmd = parse_help(PYTHON_HELP);
-        assert_eq!(cmd.name.as_deref(), Some("python3"));
-
-        let shorts = short_names(&cmd);
-        assert!(shorts.contains(&"-h"));
-        assert!(shorts.contains(&"-v"));
-        assert!(shorts.contains(&"-V"));
-        assert!(shorts.contains(&"-q"));
-
-        // Descriptions captured inline (separator is ` : `).
-        assert!(
-            cmd.args
-                .iter()
-                .find(|a| a.short.as_deref() == Some("-h"))
-                .and_then(|a| a.description.as_deref())
-                .unwrap_or("")
-                .contains("help message"),
-            "description should mention 'help message'"
-        );
-    }
-
-    // ── Python argparse example ───────────────────────────────────────────────
-
-    const ARGPARSE_EXAMPLE: &str = r#"usage: prog.py [-h] [--foo FOO] bar
-
-A sample argparse program.
-
-positional arguments:
-  bar          the bar argument
-
-optional arguments:
-  -h, --help   show this help message and exit
-  --foo FOO    the foo value
-"#;
-
-    #[test]
-    fn test_argparse_example() {
-        let cmd = parse_help(ARGPARSE_EXAMPLE);
-        assert_eq!(cmd.name.as_deref(), Some("prog.py"));
-        assert!(
-            cmd.description
-                .as_deref()
-                .unwrap_or("")
-                .contains("argparse"),
-            "description should mention argparse"
-        );
-
-        let longs = long_names(&cmd);
-        assert!(longs.contains(&"bar"));
-        assert!(longs.contains(&"--foo"));
-        assert!(longs.contains(&"--help"));
-        let shorts = short_names(&cmd);
-        assert!(shorts.contains(&"-h"));
-
-        // Inline descriptions captured by the argparse parser.
-        assert!(
-            arg_by_long(&cmd, "bar")
-                .and_then(|a| a.description.as_deref())
-                .unwrap_or("")
-                .contains("bar argument"),
-            "description should mention 'bar argument'"
-        );
-        assert!(
-            arg_by_long(&cmd, "--foo")
-                .and_then(|a| a.description.as_deref())
-                .unwrap_or("")
-                .contains("foo value"),
-            "description should mention 'foo value'"
-        );
-    }
-
-    // ── Python argparse with subparsers ──────────────────────────────────────
-
-    const ARGPARSE_SUBPARSERS: &str = r#"usage: myapp [-h] {init,run,stop} ...
-
-My Application
-
-positional arguments:
-  {init,run,stop}
-    init           Initialize the project
-    run            Run the application
-    stop           Stop the application
-
-optional arguments:
-  -h, --help       show this help message and exit
-  --verbose        Enable verbose output
-"#;
-
-    #[test]
-    fn test_argparse_subparsers() {
-        let cmd = parse_help(ARGPARSE_SUBPARSERS);
-        assert_eq!(cmd.name.as_deref(), Some("myapp"));
-        assert!(
-            cmd.description.as_deref().unwrap_or("").contains("My"),
-            "description should mention 'My'"
-        );
-
-        let longs = long_names(&cmd);
-        assert!(longs.contains(&"--help"));
-        assert!(longs.contains(&"--verbose"));
-        let shorts = short_names(&cmd);
-        assert!(shorts.contains(&"-h"));
-
-        // Inline description on --verbose.
-        assert!(
-            arg_by_long(&cmd, "--verbose")
-                .and_then(|a| a.description.as_deref())
-                .unwrap_or("")
-                .contains("verbose output"),
-            "description should mention 'verbose output'"
-        );
-    }
-
-    // ── Click (Python) ────────────────────────────────────────────────────────
-
-    const CLICK_HELP: &str = r#"Usage: cli [OPTIONS] COMMAND [ARGS]...
-
-  A simple CLI built with Click.
-
-Options:
-  --name TEXT      The person to greet.
-  --count INTEGER  Number of greetings.  [default: 1]
-  -v, --verbose    Enable verbose output.
-  --help           Show this message and exit.
-
-Commands:
-  hello    Greet someone.
-  goodbye  Say goodbye.
-"#;
-
-    #[test]
-    fn test_click_help() {
-        let cmd = parse_help(CLICK_HELP);
-        assert_eq!(cmd.name.as_deref(), Some("cli"));
-
-        let subs = subcommand_names(&cmd);
-        assert!(subs.contains(&"hello"));
-        assert!(subs.contains(&"goodbye"));
-
-        // Subcommand descriptions.
-        assert_eq!(
-            subcommand_by_name(&cmd, "hello").and_then(|s| s.description.as_deref()),
-            Some("Greet someone.")
-        );
-        assert_eq!(
-            subcommand_by_name(&cmd, "goodbye").and_then(|s| s.description.as_deref()),
-            Some("Say goodbye.")
-        );
-
-        let longs = long_names(&cmd);
-        assert!(longs.contains(&"--name"));
-        assert!(longs.contains(&"--count"));
-        assert!(longs.contains(&"--verbose"));
-        assert!(longs.contains(&"--help"));
-        let shorts = short_names(&cmd);
-        assert!(shorts.contains(&"-v"));
-    }
-
-    // ── argh (Python) ─────────────────────────────────────────────────────────
-
-    const ARGH_HELP: &str = r#"usage: myapp [-h] [--verbose] name
-
-A program using argh.
-
-positional arguments:
-  name         the name argument
-
-optional arguments:
-  -h, --help   show this help message and exit
-  --verbose    enable verbose output
-"#;
-
-    #[test]
-    fn test_argh_help() {
-        let cmd = parse_help(ARGH_HELP);
-        assert_eq!(cmd.name.as_deref(), Some("myapp"));
-        assert!(
-            cmd.description.as_deref().unwrap_or("").contains("argh"),
-            "description should mention argh"
-        );
-
-        let longs = long_names(&cmd);
-        assert!(longs.contains(&"name"));
-        assert!(longs.contains(&"--help"));
-        assert!(longs.contains(&"--verbose"));
-        let shorts = short_names(&cmd);
-        assert!(shorts.contains(&"-h"));
-
-        // Inline descriptions from argparse format.
-        assert!(
-            arg_by_long(&cmd, "name")
-                .and_then(|a| a.description.as_deref())
-                .unwrap_or("")
-                .contains("name argument"),
-            "description should mention 'name argument'"
-        );
-        assert!(
-            arg_by_long(&cmd, "--verbose")
-                .and_then(|a| a.description.as_deref())
-                .unwrap_or("")
-                .contains("verbose output"),
-            "description should mention 'verbose output'"
-        );
-    }
-
-    // ── yargs (JavaScript) ────────────────────────────────────────────────────
-
-    const YARGS_HELP: &str = r#"my-cli [options]
-
-Commands:
-  serve  Start the development server
-  build  Build the project for production
-  test   Run the test suite
-
-Options:
-      --help     Show help                 [boolean]
-      --version  Show version number       [boolean]
-  -p, --port     Port to listen on         [number]
-"#;
-
-    #[test]
-    fn test_yargs_help() {
-        let cmd = parse_help(YARGS_HELP);
-        // No "Usage:" line in this yargs output; name is not extracted.
-        let subs = subcommand_names(&cmd);
-        assert!(subs.contains(&"serve"));
-        assert!(subs.contains(&"build"));
-        assert!(subs.contains(&"test"));
-
-        // Subcommand descriptions.
-        assert!(
-            subcommand_by_name(&cmd, "serve")
-                .and_then(|s| s.description.as_deref())
-                .unwrap_or("")
-                .contains("development server"),
-            "description should mention 'development server'"
-        );
-        assert!(
-            subcommand_by_name(&cmd, "build")
-                .and_then(|s| s.description.as_deref())
-                .unwrap_or("")
-                .contains("production"),
-            "description should mention 'production'"
-        );
-
-        let longs = long_names(&cmd);
-        assert!(longs.contains(&"--help"));
-        assert!(longs.contains(&"--version"));
-        assert!(longs.contains(&"--port"));
-        let shorts = short_names(&cmd);
-        assert!(shorts.contains(&"-p"));
-    }
-
-    // ── commander (JavaScript) ────────────────────────────────────────────────
-
-    const COMMANDER_HELP: &str = r#"Usage: program [options] [command]
-
-Options:
-  -V, --version  output the version number
-  -h, --help     display help for command
-
-Commands:
-  start          Start the server
-  stop           Stop the server
-  status         Show the server status
-"#;
-
-    #[test]
-    fn test_commander_help() {
-        let cmd = parse_help(COMMANDER_HELP);
-        assert_eq!(cmd.name.as_deref(), Some("program"));
-
-        let subs = subcommand_names(&cmd);
-        assert!(subs.contains(&"start"));
-        assert!(subs.contains(&"stop"));
-        assert!(subs.contains(&"status"));
-
-        // Subcommand descriptions.
-        assert_eq!(
-            subcommand_by_name(&cmd, "start").and_then(|s| s.description.as_deref()),
-            Some("Start the server")
-        );
-        assert_eq!(
-            subcommand_by_name(&cmd, "stop").and_then(|s| s.description.as_deref()),
-            Some("Stop the server")
-        );
-
-        let longs = long_names(&cmd);
-        assert!(longs.contains(&"--version"));
-        assert!(longs.contains(&"--help"));
-        let shorts = short_names(&cmd);
-        assert!(shorts.contains(&"-V"));
-        assert!(shorts.contains(&"-h"));
-    }
-
-    // ── cobra (Go) ────────────────────────────────────────────────────────────
-
-    const COBRA_HELP: &str = r#"A brief description of your application.
-
-Usage:
-  myapp [command]
-
-Available Commands:
-  completion  Generate the autocompletion script for the specified shell
-  help        Help about any command
-  serve       Start the server
-  version     Print the version number
-
-Flags:
-  -h, --help     help for myapp
-  -t, --toggle   Help message for toggle
-
-Global Flags:
-      --config string      config file (default "$HOME/.myapp.yaml")
-      --log-level string   Log level (default "info")
-
-Use "myapp [command] --help" for more information about a command.
-"#;
-
-    #[test]
-    fn test_cobra_help() {
-        let cmd = parse_help(COBRA_HELP);
-        assert_eq!(cmd.name.as_deref(), Some("myapp"));
-        assert!(
-            cmd.description
-                .as_deref()
-                .unwrap_or("")
-                .contains("application"),
-            "description should mention application"
-        );
-
-        let subs = subcommand_names(&cmd);
-        assert!(subs.contains(&"serve"));
-        assert!(subs.contains(&"version"));
-
-        // Subcommand descriptions.
-        assert_eq!(
-            subcommand_by_name(&cmd, "serve").and_then(|s| s.description.as_deref()),
-            Some("Start the server")
-        );
-        assert!(
-            subcommand_by_name(&cmd, "completion")
-                .and_then(|s| s.description.as_deref())
-                .unwrap_or("")
-                .contains("autocompletion"),
-            "description should mention 'autocompletion'"
-        );
-
-        let longs = long_names(&cmd);
-        assert!(longs.contains(&"--help"));
-        assert!(longs.contains(&"--toggle"));
-        assert!(longs.contains(&"--config"));
-        assert!(longs.contains(&"--log-level"));
-        let shorts = short_names(&cmd);
-        assert!(shorts.contains(&"-h"));
-        assert!(shorts.contains(&"-t"));
-    }
-
-    // ── docopt (Python) ───────────────────────────────────────────────────────
-
-    const DOCOPT_HELP: &str = r#"Usage:
-  naval_fate ship <name> move <x> <y> [--speed=<kn>]
-  naval_fate ship shoot <x> <y>
-  naval_fate mine (set|remove) <x> <y> [--moored|--drifting]
-  naval_fate (-h | --help)
-  naval_fate --version
-
-Options:
-  -h --help     Show this screen.
-  --version     Show version.
-  --speed=<kn>  Speed in knots [default: 10].
-  --moored      Moored (anchored) mine.
-  --drifting    Drifting mine.
-"#;
-
-    #[test]
-    fn test_docopt_help() {
-        let cmd = parse_help(DOCOPT_HELP);
-        // Name extracted from the first usage pattern line.
-        assert_eq!(cmd.name.as_deref(), Some("naval_fate"));
-
-        let longs = long_names(&cmd);
-        assert!(longs.contains(&"--help"));
-        assert!(longs.contains(&"--version"));
-        assert!(longs.contains(&"--speed"));
-        assert!(longs.contains(&"--moored"));
-        assert!(longs.contains(&"--drifting"));
-        let shorts = short_names(&cmd);
-        assert!(shorts.contains(&"-h"));
-
-        // --speed=<kn> carries a value-type hint.
-        assert_eq!(
-            arg_by_long(&cmd, "--speed").and_then(|a| a.value_type.as_deref()),
-            Some("kn")
-        );
-    }
 
     // ── to_clap_command ───────────────────────────────────────────────────────
 
@@ -1378,424 +595,6 @@ Options:
         assert!(script.contains("readelf"));
     }
 
-    // ── git --help ───────────────────────────────────────────────────────────
-    //
-    // A simplified version of `git --help` using a standard Commands: section
-    // so that the clap parser picks up the subcommands correctly.
-
-    const GIT_HELP: &str = r#"usage: git [-v | --version] [-h | --help] [-C <path>] [-c <name>=<value>]
-           [--exec-path[=<path>]] [--html-path] [--man-path] [--info-path]
-           [--no-replace-objects] [--bare] [--git-dir=<path>] [--work-tree=<path>]
-           <command> [<args>]
-
-Usage: git [OPTIONS] <COMMAND>
-
-Commands:
-  clone     Clone a repository into a new directory
-  init      Create an empty Git repository or reinitialize an existing one
-  add       Add file contents to the index
-  rm        Remove files from the working tree and from the index
-  diff      Show changes between commits, commit and working tree, etc
-  log       Show the commit logs
-  status    Show the working tree status
-  branch    List, create, or delete branches
-  commit    Record changes to the repository
-  merge     Join two or more development histories together
-  rebase    Reapply commits on top of another base tip
-  reset     Reset current HEAD to the specified state
-  fetch     Download objects and refs from another repository
-  pull      Fetch from and integrate with another repository or a local branch
-  push      Update remote refs along with associated objects
-  tag       Create, list, delete or verify a tag object signed with GPG
-
-Options:
-  -C <path>                Run as if git was started in <path>
-  -c <name>=<value>        Pass a configuration parameter to the command
-  -v, --version            Print the Git suite version
-  -h, --help               Print help
-"#;
-
-    #[test]
-    fn test_git_help() {
-        let cmd = parse_help(GIT_HELP);
-        assert_eq!(cmd.name.as_deref(), Some("git"));
-
-        let subs = subcommand_names(&cmd);
-        assert!(subs.contains(&"clone"));
-        assert!(subs.contains(&"commit"));
-        assert!(subs.contains(&"log"));
-        assert!(subs.contains(&"diff"));
-        assert!(subs.contains(&"push"));
-        assert!(subs.contains(&"pull"));
-        assert!(subs.contains(&"branch"));
-        assert!(subs.contains(&"status"));
-
-        // Check a couple of subcommand descriptions.
-        assert_eq!(
-            subcommand_by_name(&cmd, "clone").and_then(|s| s.description.as_deref()),
-            Some("Clone a repository into a new directory")
-        );
-        assert_eq!(
-            subcommand_by_name(&cmd, "commit").and_then(|s| s.description.as_deref()),
-            Some("Record changes to the repository")
-        );
-    }
-
-    // ── git log --help ────────────────────────────────────────────────────────
-
-    const GIT_LOG_HELP: &str = r#"usage: git log [<options>] [<revision-range>] [[--] <path>...]
-
-    -q, --quiet           Suppress diff output
-    -v, --verbose         Be more verbose
-    -n, --max-count <n>   Limit the number of commits to output
-    --skip <n>            Skip number commits before starting to show output
-    --since <date>        Show commits more recent than a specific date
-    --after <date>        Show commits more recent than a specific date (alias)
-    --until <date>        Show commits older than a specific date
-    --before <date>       Show commits older than a specific date (alias)
-    --author <pattern>    Limit output to commits with matching author
-    --committer <pattern> Limit output to commits with matching committer
-    --grep <pattern>      Limit commit output to ones matching the pattern
-    --all-match           Limit output to commits matching all grep patterns
-    -i, --regexp-ignore-case
-                          Case insensitive match
-    --oneline             Show one commit per line with abbreviated hash
-    --no-walk             Only show given commits, not their parents
-    --graph               Show the commit graph as ASCII art
-    --decorate            Decorate refs
-    --first-parent        Follow only the first parent commit upon merges
-    -p, --patch           Show patch
-    -s, --no-patch        Suppress diff output
-    --stat                Generate a diffstat
-    --format <format>     Pretty-print in given format
-    --abbrev-commit       Show only a partial commit hash prefix
-"#;
-
-    #[test]
-    fn test_git_log_help() {
-        let cmd = parse_help(GIT_LOG_HELP);
-        assert_eq!(cmd.name.as_deref(), Some("git"));
-
-        let shorts = short_names(&cmd);
-        assert!(shorts.contains(&"-q"));
-        assert!(shorts.contains(&"-v"));
-        assert!(shorts.contains(&"-n"));
-        assert!(shorts.contains(&"-p"));
-        assert!(shorts.contains(&"-s"));
-        assert!(shorts.contains(&"-i"));
-
-        let longs = long_names(&cmd);
-        assert!(longs.contains(&"--max-count"));
-        assert!(longs.contains(&"--since"));
-        assert!(longs.contains(&"--author"));
-        assert!(longs.contains(&"--grep"));
-        assert!(longs.contains(&"--oneline"));
-        assert!(longs.contains(&"--graph"));
-        assert!(longs.contains(&"--stat"));
-        assert!(longs.contains(&"--first-parent"));
-        assert!(longs.contains(&"--patch"));
-        assert!(longs.contains(&"--no-patch"));
-
-        // Value types.
-        assert_eq!(
-            arg_by_long(&cmd, "--max-count").and_then(|a| a.value_type.as_deref()),
-            Some("n")
-        );
-        assert_eq!(
-            arg_by_long(&cmd, "--since").and_then(|a| a.value_type.as_deref()),
-            Some("date")
-        );
-        assert_eq!(
-            arg_by_long(&cmd, "--author").and_then(|a| a.value_type.as_deref()),
-            Some("pattern")
-        );
-
-        // Descriptions.
-        assert!(
-            arg_by_long(&cmd, "--graph")
-                .and_then(|a| a.description.as_deref())
-                .unwrap_or("")
-                .contains("ASCII"),
-            "--graph description should mention ASCII"
-        );
-    }
-
-    // ── git commit --help ────────────────────────────────────────────────────
-
-    const GIT_COMMIT_HELP: &str = r#"usage: git commit [-a | --interactive | --patch] [-s] [-v] [--amend]
-   [--author=<author>] [--date=<date>] [-m <msg>] [--no-verify]
-
-    -q, --quiet           Suppress summary after successful commit
-    -v, --verbose         Show diff in commit message template
-    -F, --file <file>     Read message from file
-    --author <author>     Override author for commit
-    --date <date>         Override date for commit
-    -m, --message <message>
-                          Commit message
-    -s, --signoff         Add a Signed-off-by trailer
-    -e, --edit            Force edit of commit
-    --cleanup <mode>      How to strip spaces and #comments from message
-    --no-verify           Bypass pre-commit and commit-msg hooks
-    --allow-empty         Allow recording an empty commit
-    --allow-empty-message
-                          Allow recording a commit with an empty message
-    --amend               Amend previous commit
-    -a, --all             Commit all changed files
-    -i, --include         Add specified files to index for commit
-    -o, --only            Commit only specified files
-    --dry-run             Show what would be committed
-    --short               Show status concisely
-    --branch              Show branch information
-    --squash <commit>     Use autosquash formatted message to squash specified commit
-    --fixup <commit>      Use autosquash formatted message to fixup specified commit
-"#;
-
-    #[test]
-    fn test_git_commit_help() {
-        let cmd = parse_help(GIT_COMMIT_HELP);
-        assert_eq!(cmd.name.as_deref(), Some("git"));
-
-        let shorts = short_names(&cmd);
-        assert!(shorts.contains(&"-q"));
-        assert!(shorts.contains(&"-v"));
-        assert!(shorts.contains(&"-m"));
-        assert!(shorts.contains(&"-a"));
-        assert!(shorts.contains(&"-s"));
-        assert!(shorts.contains(&"-e"));
-
-        let longs = long_names(&cmd);
-        assert!(longs.contains(&"--message"));
-        assert!(longs.contains(&"--author"));
-        assert!(longs.contains(&"--amend"));
-        assert!(longs.contains(&"--no-verify"));
-        assert!(longs.contains(&"--allow-empty"));
-        assert!(longs.contains(&"--dry-run"));
-        assert!(longs.contains(&"--squash"));
-        assert!(longs.contains(&"--signoff"));
-
-        // Value types.
-        assert_eq!(
-            arg_by_long(&cmd, "--message").and_then(|a| a.value_type.as_deref()),
-            Some("message")
-        );
-        assert_eq!(
-            arg_by_long(&cmd, "--file").and_then(|a| a.value_type.as_deref()),
-            Some("file")
-        );
-    }
-
-    // ── git diff --help ───────────────────────────────────────────────────────
-
-    const GIT_DIFF_HELP: &str = r#"usage: git diff [<options>] [<commit>] [--] [<path>...]
-   or: git diff [<options>] --cached [<commit>] [--] [<path>...]
-   or: git diff [<options>] <commit>...<commit> [--] [<path>...]
-
-    -p, --patch           Show patch
-    -s, --no-patch        Suppress diff output
-    --raw                 Generate the diff in raw format
-    --minimal             Produce the smallest possible diff
-    --patience            Use the patience diff algorithm
-    --histogram           Use the histogram diff algorithm
-    --diff-algorithm <algorithm>
-                          Choose a diff algorithm
-    --stat                Generate a diffstat
-    --compact-summary     Generate a compact summary in diffstat
-    --numstat             Generate a diffstat in machine-readable form
-    --shortstat           Output only the last line of --stat format
-    --name-only           Show only names of changed files
-    --name-status         Show only names and status of changed files
-    --color <when>        Show colored diff (always, never, auto)
-    --no-color            Turn off colored diff
-    --word-diff <mode>    Show a word diff using the given mode
-    -U, --unified <n>     Generate diffs with <n> lines of context
-    --cached              View the changes staged for the next commit
-    -W, --function-context
-                          Show whole surrounding functions of changes
-"#;
-
-    #[test]
-    fn test_git_diff_help() {
-        let cmd = parse_help(GIT_DIFF_HELP);
-        assert_eq!(cmd.name.as_deref(), Some("git"));
-
-        let shorts = short_names(&cmd);
-        assert!(shorts.contains(&"-p"));
-        assert!(shorts.contains(&"-s"));
-        assert!(shorts.contains(&"-U"));
-        assert!(shorts.contains(&"-W"));
-
-        let longs = long_names(&cmd);
-        assert!(longs.contains(&"--patch"));
-        assert!(longs.contains(&"--no-patch"));
-        assert!(longs.contains(&"--stat"));
-        assert!(longs.contains(&"--name-only"));
-        assert!(longs.contains(&"--cached"));
-        assert!(longs.contains(&"--word-diff"));
-        assert!(longs.contains(&"--color"));
-        assert!(longs.contains(&"--unified"));
-    }
-
-    // ── git with subcommand help synthesis (static test) ─────────────────────
-
-    #[test]
-    fn test_subcommand_synthesis_overlay() {
-        let help_runner = |args: &[&str]| -> anyhow::Result<String> {
-            let text = match args {
-                [] => GIT_HELP,
-                ["log"] => GIT_LOG_HELP,
-                ["commit"] => GIT_COMMIT_HELP,
-                ["diff"] => GIT_DIFF_HELP,
-                _ => "",
-            };
-            Ok(text.to_string())
-        };
-        let top = synthesize_completion("git", help_runner, SynthesisStrategy::RunHelp).unwrap();
-
-        assert!(subcommand_names(&top).contains(&"log"));
-        assert!(subcommand_names(&top).contains(&"commit"));
-
-        let log_sub = subcommand_by_name(&top, "log").expect("log subcommand should exist");
-        assert!(long_names(log_sub).contains(&"--oneline"));
-        assert!(long_names(log_sub).contains(&"--graph"));
-        assert!(long_names(log_sub).contains(&"--since"));
-
-        let commit_sub =
-            subcommand_by_name(&top, "commit").expect("commit subcommand should exist");
-        assert!(long_names(commit_sub).contains(&"--amend"));
-        assert!(long_names(commit_sub).contains(&"--message"));
-        assert!(long_names(commit_sub).contains(&"--no-verify"));
-
-        let diff_sub = subcommand_by_name(&top, "diff").expect("diff subcommand should exist");
-        assert!(long_names(diff_sub).contains(&"--cached"));
-        assert!(long_names(diff_sub).contains(&"--stat"));
-    }
-
-    const SIMPLE_MANPAGE: &str = r#".TH GIT-ADD 1
-.SH OPTIONS
-.TP
-.B \-p
-Interactively choose hunks.
-.TP
-.BI \-\^-intent-to-add
-Record intent-to-add without content.
-"#;
-
-    #[test]
-    fn test_synthesize_completion_run_help_strategy_uses_help_runner_only() {
-        let help_calls = Cell::new(0);
-        let manpage_calls = Cell::new(0);
-        let help_runner = |args: &[&str]| -> anyhow::Result<String> {
-            help_calls.set(help_calls.get() + 1);
-            let text = match args {
-                [] => GIT_DIFF_HELP,
-                _ => "",
-            };
-            Ok(text.to_string())
-        };
-        let manpage_loader = |_: &str| -> anyhow::Result<Command> {
-            manpage_calls.set(manpage_calls.get() + 1);
-            anyhow::bail!("manpage loader should not be used")
-        };
-
-        let cmd = synthesize_completion_with(
-            "git-add",
-            &help_runner,
-            &manpage_loader,
-            SynthesisStrategy::RunHelp,
-        )
-        .unwrap();
-
-        assert!(long_names(&cmd).contains(&"--patch"));
-        assert_eq!(help_calls.get(), 1);
-        assert_eq!(manpage_calls.get(), 0);
-    }
-
-    #[test]
-    fn test_synthesize_completion_man_page_strategy_uses_manpage_only() {
-        let help_calls = Cell::new(0);
-        let manpage_calls = Cell::new(0);
-        let help_runner = |_: &[&str]| -> anyhow::Result<String> {
-            help_calls.set(help_calls.get() + 1);
-            anyhow::bail!("help runner should not be used")
-        };
-        let manpage_loader = |command_path: &str| -> anyhow::Result<Command> {
-            manpage_calls.set(manpage_calls.get() + 1);
-            assert_eq!(command_path, "git-add");
-            Ok(parse_man::parse_manpage("git-add", SIMPLE_MANPAGE).unwrap())
-        };
-
-        let cmd = synthesize_completion_with(
-            "git-add",
-            &help_runner,
-            &manpage_loader,
-            SynthesisStrategy::ManPage,
-        )
-        .unwrap();
-
-        assert_eq!(cmd.name.as_deref(), Some("git-add"));
-        assert!(short_names(&cmd).contains(&"-p"));
-        assert!(long_names(&cmd).contains(&"--intent-to-add"));
-        assert_eq!(help_calls.get(), 0);
-        assert_eq!(manpage_calls.get(), 1);
-    }
-
-    #[test]
-    fn test_synthesize_completion_man_page_then_run_help_falls_back() {
-        let help_calls = Cell::new(0);
-        let manpage_calls = Cell::new(0);
-        let help_runner = |args: &[&str]| -> anyhow::Result<String> {
-            help_calls.set(help_calls.get() + 1);
-            let text = match args {
-                [] => GIT_DIFF_HELP,
-                _ => "",
-            };
-            Ok(text.to_string())
-        };
-        let manpage_loader = |_: &str| -> anyhow::Result<Command> {
-            manpage_calls.set(manpage_calls.get() + 1);
-            anyhow::bail!("missing man page")
-        };
-
-        let cmd = synthesize_completion_with(
-            "git-add",
-            &help_runner,
-            &manpage_loader,
-            SynthesisStrategy::ManPageThenRunHelp,
-        )
-        .unwrap();
-
-        assert!(long_names(&cmd).contains(&"--patch"));
-        assert_eq!(help_calls.get(), 1);
-        assert_eq!(manpage_calls.get(), 1);
-    }
-
-    #[test]
-    fn test_synthesize_completion_man_page_then_run_help_prefers_manpage() {
-        let help_calls = Cell::new(0);
-        let manpage_calls = Cell::new(0);
-        let help_runner = |_: &[&str]| -> anyhow::Result<String> {
-            help_calls.set(help_calls.get() + 1);
-            Ok(GIT_DIFF_HELP.to_string())
-        };
-        let manpage_loader = |_: &str| -> anyhow::Result<Command> {
-            manpage_calls.set(manpage_calls.get() + 1);
-            Ok(parse_man::parse_manpage("git-add", SIMPLE_MANPAGE).unwrap())
-        };
-
-        let cmd = synthesize_completion_with(
-            "git-add",
-            &help_runner,
-            &manpage_loader,
-            SynthesisStrategy::ManPageThenRunHelp,
-        )
-        .unwrap();
-
-        assert!(long_names(&cmd).contains(&"--intent-to-add"));
-        assert!(!long_names(&cmd).contains(&"--patch"));
-        assert_eq!(help_calls.get(), 0);
-        assert_eq!(manpage_calls.get(), 1);
-    }
-
     // ── find_subcommand_mut ───────────────────────────────────────────────────
 
     #[test]
@@ -1857,195 +656,5 @@ Record intent-to-add without content.
         // A path that doesn't exist should return None.
         let path = vec!["nonexistent".to_string()];
         assert!(find_subcommand_mut(&mut root, &path).is_none());
-    }
-
-    // ── Nested subcommand synthesis (static simulation) ───────────────────────
-    //
-    // Verify the stack-based iteration finds sub-sub-commands.  We simulate
-    // a two-level subcommand tree (remote → add/set-url) and confirm that
-    // args are populated at both levels.
-
-    const GIT_REMOTE_HELP: &str = r#"Usage: git remote [OPTIONS] <COMMAND>
-
-Commands:
-  add      Add a remote
-  set-url  Changes URLs for the remote
-  remove   Remove a remote
-  rename   Rename a remote
-
-Options:
-  -v, --verbose    Be verbose
-  -h, --help       Print help
-"#;
-
-    const GIT_REMOTE_ADD_HELP: &str = r#"usage: git remote add <name> <url>
-
-    -f, --fetch        Fetch after adding the remote
-    -t, --track <branch>
-                       Track given branch
-    --tags             Import all tags
-    --no-tags          Do not import tags
-"#;
-
-    const GIT_REMOTE_SET_URL_HELP: &str = r#"usage: git remote set-url [--push] <name> <new-url>
-
-    --push             Manipulate push URLs
-    --add              Add URL to list
-    --delete           Delete URL
-"#;
-
-    const GIT_WITH_REMOTE_HELP: &str = r#"Usage: git [OPTIONS] <COMMAND>
-
-Commands:
-  clone   Clone a repository into a new directory
-  remote  Manage set of tracked repositories
-
-Options:
-  -h, --help  Print help
-"#;
-
-    #[test]
-    fn test_nested_subcommand_synthesis_simulation() {
-        let help_runner = |args: &[&str]| -> anyhow::Result<String> {
-            let text = match args {
-                [] => GIT_WITH_REMOTE_HELP,
-                ["remote"] => GIT_REMOTE_HELP,
-                ["remote", "add"] => GIT_REMOTE_ADD_HELP,
-                ["remote", "set-url"] => GIT_REMOTE_SET_URL_HELP,
-                _ => "",
-            };
-            Ok(text.to_string())
-        };
-        let top = synthesize_completion("git", help_runner, SynthesisStrategy::RunHelp).unwrap();
-
-        let remote =
-            subcommand_by_name(&top, "remote").expect("remote subcommand should be present");
-        assert!(long_names(remote).contains(&"--verbose"));
-
-        let remote_add =
-            subcommand_by_name(remote, "add").expect("remote add subcommand should be present");
-        assert!(long_names(remote_add).contains(&"--fetch"));
-        assert!(short_names(remote_add).contains(&"-f"));
-        assert_eq!(
-            arg_by_long(remote_add, "--track").and_then(|a| a.value_type.as_deref()),
-            Some("branch")
-        );
-
-        let remote_set_url = subcommand_by_name(remote, "set-url")
-            .expect("remote set-url subcommand should be present");
-        assert!(long_names(remote_set_url).contains(&"--push"));
-        assert!(long_names(remote_set_url).contains(&"--delete"));
-    }
-
-    // ── readelf --help ────────────────────────────────────────────────────────
-    //
-    // readelf uses a non-standard "Options are:" section header and inline
-    // descriptions, making it a useful real-world test for the generic
-    // section-header recognition and the inline-description capture paths.
-
-    const READELF_HELP: &str = r#"Usage: readelf <option(s)> elf-file(s)
- Display information about the contents of ELF format files
- Options are:
-  -a --all               Equivalent to: -h -l -S -s -r -d -V -A -I
-  -h --file-header       Display the ELF file header
-  -l --program-headers   Display the program headers
-     --segments          An alias for --program-headers
-  -S --section-headers   Display the sections' header
-     --sections          An alias for --section-headers
-  -g --section-groups    Display the section groups
-  -t --section-details   Display the section details
-  -e --headers           Equivalent to: -h -l -S
-  -s --syms              Display the symbol table
-     --symbols           An alias for --syms
-     --dyn-syms          Display the dynamic symbol table
-     --lto-syms          Display LTO symbol tables
-     --sym-base=[0|8|10|16] 
-                         Force base for symbol sizes.  The options are 
-                         mixed (the default), octal, decimal, hexadecimal.
-  -C --demangle[=STYLE]  Decode mangled/processed symbol names
-                           STYLE can be "none", "auto", "gnu-v3", "java",
-                           "gnat", "dlang", "rust"
-     --no-demangle       Do not demangle low-level symbol names.  (default)
-     --recurse-limit     Enable a demangling recursion limit.  (default)
-     --no-recurse-limit  Disable a demangling recursion limit
-  -n --notes             Display the contents of note sections (if present)
-  -r --relocs            Display the relocations (if present)
-  -u --unwind            Display the unwind info (if present)
-  -d --dynamic           Display the dynamic section (if present)
-  -V --version-info      Display the version sections (if present)
-  -A --arch-specific     Display architecture specific information (if any)
-  -c --archive-index     Display the symbol/file index in an archive
-  -D --use-dynamic       Use the dynamic section info when displaying symbols
-  -L --lint|--enable-checks
-                         Display warning messages for possible problems
-  -x --hex-dump=<number|name>
-                         Dump the contents of section <number|name> as bytes
-  -p --string-dump=<number|name>
-                         Dump the contents of section <number|name> as strings
-  -R --relocated-dump=<number|name>
-                         Dump the relocated contents of section <number|name>
-  -z --decompress        Decompress section before dumping it
-  -I --histogram         Display histogram of bucket list lengths
-  -W --wide              Allow output width to exceed 80 characters
-  -T --silent-truncation If a symbol name is truncated, do not add [...] suffix
-  -H --help              Display this information
-  -v --version           Display the version number of readelf
-Report bugs to <https://sourceware.org/bugzilla/>
-"#;
-
-    #[test]
-    fn test_readelf_help() {
-        let cmd = parse_help(READELF_HELP);
-        assert_eq!(cmd.name.as_deref(), Some("readelf"));
-
-        // Key short flags.
-        let shorts = short_names(&cmd);
-        assert!(shorts.contains(&"-a"));
-        assert!(shorts.contains(&"-h"));
-        assert!(shorts.contains(&"-l"));
-        assert!(shorts.contains(&"-S"));
-        assert!(shorts.contains(&"-r"));
-        assert!(shorts.contains(&"-d"));
-        assert!(shorts.contains(&"-n"));
-        assert!(shorts.contains(&"-W"));
-        assert!(shorts.contains(&"-H"));
-        assert!(shorts.contains(&"-v"));
-
-        // Key long flags.
-        let longs = long_names(&cmd);
-        assert!(longs.contains(&"--all"));
-        assert!(longs.contains(&"--file-header"));
-        assert!(longs.contains(&"--program-headers"));
-        assert!(longs.contains(&"--section-headers"));
-        assert!(longs.contains(&"--relocs"));
-        assert!(longs.contains(&"--dynamic"));
-        assert!(longs.contains(&"--notes"));
-        assert!(longs.contains(&"--wide"));
-        assert!(longs.contains(&"--help"));
-        assert!(longs.contains(&"--version"));
-        assert!(longs.contains(&"--hex-dump"));
-
-        // Value type extracted from --hex-dump=<number|name>.
-        assert_eq!(
-            arg_by_long(&cmd, "--hex-dump").and_then(|a| a.value_type.as_deref()),
-            Some("number|name")
-        );
-    }
-
-    #[test]
-    fn test_parse_flag_tokens_ignores_multi_character_short_forms() {
-        let (short, long, value_type) = parse_flag_tokens("-wk --debug-dump=links");
-        assert_eq!(short, None);
-        assert_eq!(long.as_deref(), Some("--debug-dump"));
-        assert_eq!(value_type.as_deref(), Some("links"));
-
-        let (short, long, value_type) =
-            parse_flag_tokens("-U[dlexhi] --unicode=[default|locale|escape|hex|highlight|invalid]");
-        assert_eq!(short, None);
-        assert_eq!(long.as_deref(), Some("--unicode"));
-        assert_eq!(
-            value_type.as_deref(),
-            Some("[default|locale|escape|hex|highlight|invalid]")
-        );
     }
 }
